+++ CHANGES.md
@@ -5,6 +5,8 @@ Markdown Writer FX Change Log
 
 - Move selected lines up or down with `Alt+Up` or `Alt+Down`.
 - Duplicate selected lines up or down with `Ctrl+Alt+Up` or `Ctrl+Alt+Down`.
+- Indent or unindent lines with `Tab` or `Shift+Tab`. `Backspace` key unindents
+  line if caret is in leading whitespace of a line.
 - Smart "Insert > Header X": Support toggle header and change header level.
 - Smarter "Insert > Bold/Italic/Strikethrough/Inline Code": if selection contains
   bold/italic/strikethrough/inline-code text, then it will be changed to plain text.